+++ CHANGELOG.md
@@ -9,6 +9,10 @@ v0.9.14-dev (January xx, 2022)
 
 * Change label 'certified statement' to 'reference statement'.
 
+**Bug fixes**
+
+* Footer fixes.
+
 
 v0.9.13 (January 23, 2022)
 --------------------------