+++ CHANGELOG.md
@@ -13,6 +13,11 @@ v0.9.14-dev (January xx, 2022)
 
 * Add support for OIDC SSO configuration separate from OKTA SSO configuration.
 
+**Bug fixes**
+
+* Footer fixes.
+
+
 v0.9.13 (January 23, 2022)
 --------------------------