+++ CHANGELOG.md
@@ -12,6 +12,7 @@ v0.9.14-dev (January xx, 2022)
 **Developer changes**
 
 * Add support for OIDC SSO configuration separate from OKTA SSO configuration.
+* Update Django, libraries.
 
 **Bug fixes**